Moved syntax highlighting to its own file
